Time: 16 ms (73.26%), Space: 144.8 MB (95.19%) - LeetHub

+++ 2529-range-product-queries-of-powers/2529-range-product-queries-of-powers.cpp
@@ -1,16 +1,16 @@
 class Solution {
 public:
     vector<int> productQueries(int n, vector<vector<int>>& queries) {
-        vector<long long> powers;
+        vector<int> powers;
         int sum=0;
         for(int i=0;i<31;i++){
-            if( n & (1ll<<i) ){
-                powers.push_back(1ll<<i);
+            if( n & (1<<i) ){
+                powers.push_back(1<<i);
             }
         }
         vector<int> ans(queries.size());
         for(int i=0;i<queries.size();i++){
-            long long prod=1;
+            int prod=1;
             for(int j=queries[i][0];j<=queries[i][1];j++){
                 int mod=1e9+7;
                 prod=(1ll*prod*powers[j])%(mod);
